requirements-dev.txt
-r ./requirements.txt
black
isort
ruff
coverage
sphinx~=4.4.0
myst-parser~=0.17.0
furo~=2022.2
pre-commit
django-debug-toolbar~=3.2.4
fonttools~=4.33.3
pytest
pytest-django